Mingyuanm/sdxl export (#8926)
* Move cached embedding devices and dtype for onnx export consistency


* Add old trt export/inference script, currently not working in latest container.


* Add NeMo TRT inference pipeline and quatization workflow


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add guards to avoid undefined variables


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* minor fix


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Add conversion script from hf sdxl to nemo sdxl


* Update quantize pipeline to adapt to variable image dimension


* update sdxl pipeline to be aware of additional emb channels


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* add guards for potential local var


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* copyright header


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update calib prompt file path


* Update file paths


* minor update


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update default quantization config


* remove unused imports/vars


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* remove unused imports


---------

+++ nemo/collections/multimodal/modules/stable_diffusion/diffusionmodules/util.py
@@ -207,7 +207,7 @@ def timestep_embedding(timesteps, dim, max_period=10000, repeat_only=False, cach
     if not repeat_only:
         if cached_embedding is not None:
             # using cached embedding and lookup in the cache
-            embedding = cached_embedding[timesteps, :]
+            embedding = cached_embedding[timesteps.to(dtype=torch.int), :]
         else:
             half = dim // 2
             idx = get_idx(half, timesteps.device)
